% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{browse_templates}
\alias{browse_templates}
\title{Open master wikis in web browser for a given set of rally OSF ids}
\usage{
browse_templates(rally_ids, edit = TRUE)
}
\arguments{
\item{rally_ids}{vector of OSF component ids pointing to indivual rally spaces}

\item{edit}{should the pages be opened in edit view?}
}
\description{
Open master wikis in web browser for a given set of rally OSF ids
}
