% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download file from OSF}
\usage{
download_file(id, dest, pat = get_osf_pat())
}
\arguments{
\item{id}{OSF ID of file.}

\item{dest}{destination to place downloaded file.}

\item{pat}{pat OSF personal access token.}
}
\description{
Download file from OSF
}
