% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{gen_questions_data}
\alias{gen_questions_data}
\title{Generate questions page data from rally content}
\usage{
gen_questions_data(content, base_path = get_rally_base_path())
}
\arguments{
\item{content}{list of rally template content obtained from \code{\link{get_rally_content}}}

\item{base_path}{location of base path where outputs should be stored}
}
\description{
Generate questions page data from rally content
}
