% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest}
\alias{get_latest}
\title{Get latest rally data from cache (compared to supplied input object)}
\usage{
get_latest(obj)
}
\arguments{
\item{obj}{an element of the list returned from \code{\link{get_rally_content}}}
}
\description{
Get latest rally data from cache (compared to supplied input object)
}
