% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latests}
\alias{get_latests}
\title{Get latest rally data from cache (compared to supplied input objects)}
\usage{
get_latests(objs)
}
\arguments{
\item{objs}{a list of rally output objects obtained from \code{\link{get_rally_content}}}
}
\description{
Get latest rally data from cache (compared to supplied input objects)
}
