% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_osf_pat}
\alias{get_osf_pat}
\title{Retrieve OSF personal access token environment variable OSF_PAT}
\usage{
get_osf_pat()
}
\description{
Retrieve OSF personal access token environment variable OSF_PAT
}
