% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rally_base_path}
\alias{get_rally_base_path}
\title{Retrieve rally base path environment variable RALLY_BASE_PATH}
\usage{
get_rally_base_path()
}
\description{
Retrieve rally base path environment variable RALLY_BASE_PATH
}
