% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{get_rally_groups}
\alias{get_rally_groups}
\title{Get rally group information}
\usage{
get_rally_groups(root_id = "s7p4z")
}
\arguments{
\item{root_id}{root rally OSF component id}
}
\description{
Get rally group information
}
