% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{gen_overview_data}
\alias{gen_overview_data}
\title{Generate data for overview pages}
\usage{
gen_overview_data(content, base_path = get_rally_base_path())
}
\arguments{
\item{content}{list of rally template content obtained from \code{\link{get_rally_content}}}

\item{base_path}{location of base path where outputs should be stored}
}
\description{
Generate data for overview pages
}
