% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{get_rally_content}
\alias{get_rally_content}
\title{Get and parse full rally template data for a either a base rally OSF id or a vector of rally OSF ids}
\usage{
get_rally_content(root_id = NULL, rally_ids = NULL, force = FALSE)
}
\arguments{
\item{root_id}{root rally OSF component id}

\item{rally_ids}{vector of OSF component ids pointing to indivual rally spaces}

\item{force}{should template content be forced to be read and parsed from OSF instead of from the cache even if text content matches?}
}
\description{
Get and parse full rally template data for a either a base rally OSF id or a vector of rally OSF ids
}
\examples{
\dontrun{
# init_rally_engine() # one time only
rally_ids <- get_rally_ids()
content <- get_rally_content(rally_ids = rally_ids)
gen_overview_data(content)
gen_dashboard_data(content)
gen_questions_data(content)
gen_ppt(content)
}
}
